src/conway_polynomials/CPimport.txt: add the latest data file
First commit to the repository. Add the upstream data file that is the
heart of this project. We need this in the repository because we need
to ship it to users, but it will also be nice to be able to diff the
database in the future since there's no version or change information
on the website.