COPYING,LICENSE: include GPLv3+ license text with the package
If you're redistributing this package or a derivative work, the GPLv3
says that you should include a copy of the license along with it. To
make life easier for packagers and developers, we include the text of
the GPLv3 in a new LICENSE file. A new COPYING file then elaborates
that you can use either the GPLv3, or any later version.